pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.48.0",
    "mcp>=1.3.0",
    "openai>=1.65.1",
    "python-dotenv>=1.0.1",
]
